fix(PodcastToShorts): fixed matching end_text from llama response
